fix : 로컬 로그인 시 세션 미저장 수정
