+++ src/main/java/com/dokdok/user/dto/request/DevLoginRequest.java
@@ -1,0 +1,3 @@
+package com.dokdok.user.dto.request;
+
+public record DevLoginRequest(String loginId, String password) {}
